ayurvedacare-healthgpt-rag-main/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
groq>=0.5.0
chromadb>=0.4.24
ollama>=0.1.8
PyPDF2==3.0.1
python-docx>=1.1.0
sentence-transformers>=2.7.0
numpy>=1.26.4
requests==2.31.0
setuptools>=68.0.0
wheel
